Refactor LeaveCommand to minimise duplicated code
