+++ src/main/java/io/github/dinty1/easychannels/command/AbstractChannelOrientedCommand.java
@@ -1,0 +1,46 @@
+package io.github.dinty1.easychannels.command;
+
+import io.github.dinty1.easychannels.EasyChannels;
+import io.github.dinty1.easychannels.object.Channel;
+import org.bukkit.ChatColor;
+import org.bukkit.command.Command;
+import org.bukkit.command.CommandExecutor;
+import org.bukkit.command.CommandSender;
+import org.bukkit.command.TabExecutor;
+import org.bukkit.entity.Player;
+import org.jetbrains.annotations.NotNull;
+import org.jetbrains.annotations.Nullable;
+
+import java.util.List;
+
+public abstract class AbstractChannelOrientedCommand implements CommandExecutor, TabExecutor {
+
+    public abstract void execute(Channel channel, Player player);
+
+    @Override
+    public boolean onCommand(@NotNull CommandSender sender, @NotNull Command command, @NotNull String s, @NotNull String[] args) {
+        if (args.length < 1) return false;
+        Channel channel = EasyChannels.getChannelManager().getChannel(args[0]);
+        if (channel == null) return false;
+
+        if (!(sender instanceof Player)) {
+            sender.sendMessage(ChatColor.RED + "Only players can use this command.");
+            return true;
+        }
+        Player player = (Player) sender;
+
+        if (channel.getPermission() != null && !player.hasPermission(channel.getPermission()))
+            player.sendMessage(ChatColor.RED + "You do not have access to that channel.");
+
+        else this.execute(channel, player);
+
+        return true;
+    }
+
+
+    @Nullable
+    @Override
+    public List<String> onTabComplete(@NotNull CommandSender sender, @NotNull Command command, @NotNull String s, @NotNull String[] args) {
+        return EasyChannels.getChannelManager().getChannelList();
+    }
+}
